native: const-correctness + remaining long->s32 (group C) — PHASE-1 COMPLETE 573/573
GCC enforces const/qualifier rules MWcc skipped:
- riccohook: bind polarXZ() rvalue to a value, not a non-const ref.
- smallEnemy: drop 'volatile' on local_14 (codegen fakematch quirk).
- JAIGFrameSe: cast const Vec* -> Vec* for the read-only MTXMultVec src (SDK API
  isn't const-correct).
- JKRFileFinder: mFileName -> const char* (only ever read as a string).
- MActorData: cast const strstr/strrchr(param_1) results into the reused char*.
- MapObjHide: wrapName -> const char* (strstr of const mName).
- MapObjTrap: cast const TBGCheckData* (SMS_GetMarioGrPlane) to mutable.
- MapObjBase: emitAnd{Rotate,Scale,BindScale,RotateScale,SRT} long->s32.

All 573 non-asm .cpp now compile-clean (g++ -std=c++17 -m64 -Wno-narrowing,
shims force-included). The 32 asm-token files are GC OS/CRT/HW the platform owns.

+++ include/MoveBG/MapObjBase.hpp
@@ -208,19 +208,19 @@ public:
 	void initUnique();
 	void setMatTableTex(J3DMaterialTable*);
 	void setMatTable(J3DMaterialTable*);
-	JPABaseEmitter* emitAndRotate(long, unsigned char,
+	JPABaseEmitter* emitAndRotate(s32, unsigned char,
 	                              const JGeometry::TVec3<f32>*) const;
-	JPABaseEmitter* emitAndScale(long, unsigned char,
+	JPABaseEmitter* emitAndScale(s32, unsigned char,
 	                             const JGeometry::TVec3<f32>*,
 	                             const JGeometry::TVec3<f32>&) const;
-	JPABaseEmitter* emitAndBindScale(long, unsigned char,
+	JPABaseEmitter* emitAndBindScale(s32, unsigned char,
 	                                 const JGeometry::TVec3<f32>*,
 	                                 const JGeometry::TVec3<f32>&) const;
-	JPABaseEmitter* emitAndScale(long, unsigned char,
+	JPABaseEmitter* emitAndScale(s32, unsigned char,
 	                             const JGeometry::TVec3<f32>*) const;
-	JPABaseEmitter* emitAndRotateScale(long, unsigned char,
+	JPABaseEmitter* emitAndRotateScale(s32, unsigned char,
 	                                   const JGeometry::TVec3<f32>*) const;
-	JPABaseEmitter* emitAndSRT(long, unsigned char,
+	JPABaseEmitter* emitAndSRT(s32, unsigned char,
 	                           const JGeometry::TVec3<f32>*,
 	                           const JGeometry::TVec3<f32>&,
 	                           const JGeometry::TVec3<f32>&);
